+++ xtools/base.py
@@ -41,7 +41,8 @@ def error_exception(response: dict) -> Optional[BaseXToolsException]:
         return BaseXToolsException(error["message"], code)
 
     if error.startswith("No matching result") \
-            or error.endswith("is not a valid project"):
+            or error.endswith("is not a valid project") \
+            or error == "The requested user does not exist":
         return NotFound(error)
 
     if error.startswith("User has made too many edits!"):